Put all tests that compare to Brian Hargreaves' code in one file.
