+++ test/runtests.jl
@@ -6,13 +6,15 @@ using Statistics
 using Test
 
 @testset "BlochSim.jl" begin
-    include("blochmatrix.jl")
-    include("AbstractSpin.jl")
-    include("AbstractSpoiling.jl")
-    include("helpers.jl")
-    include("freeprecess.jl")
-    include("excite.jl")
-    include("expm.jl")
-    include("Spin.jl")
-    include("sequences.jl")
+#    include("blochmatrix.jl")
+#    include("AbstractSpin.jl")
+#    include("AbstractSpoiling.jl")
+#    include("helpers.jl")
+#    include("freeprecess.jl")
+#    include("excite.jl")
+#    include("expm.jl")
+#    include("Spin.jl")
+#    include("sequences.jl")
+
+    include("matlab.jl")
 end
